+++ audiobook/src/main/kotlin/de/ph1b/audiobook/dialog/JumpToPositionDialogFragment.kt
@@ -27,20 +27,22 @@ import de.ph1b.audiobook.R
 import de.ph1b.audiobook.injection.App
 import de.ph1b.audiobook.persistence.PrefsManager
 import de.ph1b.audiobook.playback.MediaPlayerController
-import de.ph1b.audiobook.uitools.ThemeUtil
+import de.ph1b.audiobook.uitools.theme
 import de.ph1b.audiobook.utils.BookVendor
 import java.util.concurrent.TimeUnit
 import javax.inject.Inject
 
 class JumpToPositionDialogFragment : DialogFragment() {
 
+    init {
+        App.component().inject(this)
+    }
+
     @Inject internal lateinit var prefs: PrefsManager
     @Inject internal lateinit var bookVendor: BookVendor
     @Inject internal lateinit var mediaPlayerController: MediaPlayerController
 
     override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
-        App.component().inject(this)
-
         // find views
         val v = activity.layoutInflater.inflate(R.layout.dialog_time_picker, null)
         val mPicker = v.findViewById(R.id.number_minute) as NumberPicker
@@ -94,8 +96,8 @@ class JumpToPositionDialogFragment : DialogFragment() {
             }
         }
 
-        ThemeUtil.theme(mPicker)
-        ThemeUtil.theme(hPicker)
+        mPicker.theme()
+        hPicker.theme()
 
         return MaterialDialog.Builder(context)
                 .customView(v, true)