+++ audiobook/src/main/kotlin/de/ph1b/audiobook/fragment/BookPlayFragment.kt
@@ -17,9 +17,9 @@
 
 package de.ph1b.audiobook.fragment
 
-import android.content.Context
 import android.content.Intent
 import android.os.Bundle
+import android.support.v4.app.Fragment
 import android.support.v4.content.ContextCompat
 import android.support.v4.view.ViewCompat
 import android.support.v7.app.AppCompatActivity
@@ -29,6 +29,7 @@ import android.widget.ImageView
 import android.widget.SeekBar
 import android.widget.Spinner
 import android.widget.TextView
+import butterknife.bindView
 import com.getbase.floatingactionbutton.FloatingActionButton
 import com.jakewharton.rxbinding.view.clicks
 import com.jakewharton.rxbinding.widget.RxAdapterView
@@ -37,6 +38,7 @@ import com.jakewharton.rxbinding.widget.SeekBarProgressChangeEvent
 import com.jakewharton.rxbinding.widget.SeekBarStopChangeEvent
 import com.squareup.picasso.Picasso
 import de.ph1b.audiobook.R
+import de.ph1b.audiobook.actionBar
 import de.ph1b.audiobook.activity.SettingsActivity
 import de.ph1b.audiobook.adapter.MultiLineSpinnerAdapter
 import de.ph1b.audiobook.dialog.BookmarkDialogFragment
@@ -46,9 +48,9 @@ import de.ph1b.audiobook.injection.App
 import de.ph1b.audiobook.model.Book
 import de.ph1b.audiobook.persistence.BookChest
 import de.ph1b.audiobook.persistence.PrefsManager
+import de.ph1b.audiobook.playback.MediaPlayerCapabilities
 import de.ph1b.audiobook.playback.MediaPlayerController
 import de.ph1b.audiobook.playback.PlayStateManager
-import de.ph1b.audiobook.playback.Player
 import de.ph1b.audiobook.uitools.CoverReplacement
 import de.ph1b.audiobook.uitools.PlayPauseDrawable
 import de.ph1b.audiobook.uitools.ThemeUtil
@@ -68,25 +70,37 @@ import javax.inject.Inject
 
  * @author Paul Woitaschek
  */
-class BookPlayFragment : BaseFragment() {
+class BookPlayFragment : Fragment() {
+
+    init {
+        App.component().inject(this)
+    }
 
     @Inject internal lateinit var mediaPlayerController: MediaPlayerController
     @Inject internal lateinit var prefs: PrefsManager
     @Inject internal lateinit var db: BookChest
     @Inject internal lateinit var bookVendor: BookVendor
     @Inject internal lateinit var playStateManager: PlayStateManager
+    @Inject internal lateinit var playerCapabilities: MediaPlayerCapabilities
 
     private val playPauseDrawable = PlayPauseDrawable()
     private var subscriptions: CompositeSubscription? = null
     private var book: Book? = null
 
-    private lateinit var hostingActivity: AppCompatActivity
+    private val hostingActivity: AppCompatActivity by lazy { context as AppCompatActivity }
 
-    private lateinit var timerCountdownView: TextView
-    private lateinit var playedTimeView: TextView
-    private lateinit var seekBar: SeekBar
-    private lateinit var bookSpinner: Spinner
-    private lateinit var maxTimeView: TextView
+    private val timerCountdownView: TextView by bindView(R.id.timerView)
+    private val playedTimeView: TextView by bindView(R.id.played)
+    private val seekBar: SeekBar by bindView(R.id.seekBar)
+    private val bookSpinner: Spinner by bindView(R.id.book_spinner)
+    private val maxTimeView: TextView by bindView(R.id.maxTime)
+    private val coverFrame: View by bindView(R.id.cover_frame)
+    private val coverView: ImageView by bindView(R.id.book_cover)
+    private val nextButton: View  by bindView(R.id.next)
+    private val fastForwardButton: View by bindView(R.id.fastForward)
+    private val playButton: FloatingActionButton by bindView(R.id.play)
+    private val rewindButton: View by bindView(R.id.rewind)
+    private val previousButton: View by bindView(R.id.previous)
 
     /**
      * @return the book id this fragment was instantiated with.
@@ -94,22 +108,8 @@ class BookPlayFragment : BaseFragment() {
     val bookId: Long
         get() = arguments.getLong(NI_BOOK_ID)
 
-    private lateinit var coverFrame: View
-
-    override fun onCreateView(inflater: LayoutInflater?, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        val view = inflater!!.inflate(R.layout.fragment_book_play, container, false)
-        timerCountdownView = view.findViewById(R.id.timerView) as TextView
-        maxTimeView = view.findViewById(R.id.maxTime) as TextView
-        bookSpinner = view.findViewById(R.id.book_spinner) as Spinner
-        seekBar = view.findViewById(R.id.seekBar) as SeekBar
-        playedTimeView = view.findViewById(R.id.played) as TextView
-        coverFrame = view.findViewById(R.id.cover_frame)
-        val coverView = view.findViewById(R.id.book_cover) as ImageView
-        val nextButton = view.findViewById(R.id.next)
-        val fastForwardButton = view.findViewById(R.id.fastForward)
-        val playButton = view.findViewById(R.id.play) as FloatingActionButton
-        val rewindButton = view.findViewById(R.id.rewind)
-        val previousButton = view.findViewById(R.id.previous)
+    override fun onViewCreated(view: View?, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
 
         playButton.clicks()
                 .onBackpressureLatest()
@@ -145,7 +145,10 @@ class BookPlayFragment : BaseFragment() {
         book = bookVendor.byId(bookId)
 
         //init views
-        hostingActivity.supportActionBar.setDisplayHomeAsUpEnabled(true)
+        val actionBar = actionBar().apply {
+            setDisplayHomeAsUpEnabled(true)
+            setHomeAsUpIndicator(R.drawable.abc_ic_ab_back_mtrl_am_alpha)
+        }
 
         //setup buttons
         playButton.setIconDrawable(playPauseDrawable)
@@ -165,7 +168,7 @@ class BookPlayFragment : BaseFragment() {
                 }
 
         if (book != null) {
-            hostingActivity.supportActionBar.title = book!!.name
+            actionBar.title = book!!.name
 
             // adapter
             val chapters = book!!.chapters
@@ -235,14 +238,13 @@ class BookPlayFragment : BaseFragment() {
                 }
             })
         }
-
-        return view
     }
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?) =
+            inflater.inflate(R.layout.fragment_book_play, container, false)
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
-        App.component().inject(this)
 
         setHasOptionsMenu(true)
     }
@@ -252,17 +254,11 @@ class BookPlayFragment : BaseFragment() {
         JumpToPositionDialogFragment().show(fragmentManager, JumpToPositionDialogFragment.TAG)
     }
 
-    override fun onAttach(context: Context?) {
-        super.onAttach(context)
-
-        hostingActivity = context as AppCompatActivity
-    }
-
     override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
         inflater.inflate(R.menu.book_play, menu)
 
         val speedItem = menu.findItem(R.id.action_time_lapse)
-        speedItem.setEnabled(Player.canSetSpeed)
+        speedItem.isEnabled = playerCapabilities.useCustomMediaPlayer
 
         // sets the correct sleep timer icon
         val sleepTimerItem = menu.findItem(R.id.action_sleep)
@@ -276,12 +272,12 @@ class BookPlayFragment : BaseFragment() {
         val currentBookExists = book != null
         val bookmarkItem = menu.findItem(R.id.action_bookmark)
         val timeChangeItem = menu.findItem(R.id.action_time_change)
-        bookmarkItem.setVisible(currentBookExists)
-        timeChangeItem.setVisible(currentBookExists)
+        bookmarkItem.isVisible = currentBookExists
+        timeChangeItem.isVisible = currentBookExists
     }
 
-    override fun onOptionsItemSelected(item: MenuItem?): Boolean {
-        when (item!!.itemId) {
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        when (item.itemId) {
             R.id.action_settings -> {
                 startActivity(Intent(context, SettingsActivity::class.java))
                 return true
@@ -321,7 +317,6 @@ class BookPlayFragment : BaseFragment() {
 
         subscriptions = CompositeSubscription().apply {
             add(playStateManager.playState
-                    .observeOn(AndroidSchedulers.mainThread())
                     .subscribe(object : Action1<PlayStateManager.PlayState> {
                         private var firstRun = true
 
@@ -425,14 +420,10 @@ class BookPlayFragment : BaseFragment() {
          * *
          * @return The new instance
          */
-        fun newInstance(bookId: Long): BookPlayFragment {
-            val bookPlayFragment = BookPlayFragment()
-
-            val args = Bundle()
-            args.putLong(NI_BOOK_ID, bookId)
-            bookPlayFragment.arguments = args
-
-            return bookPlayFragment
+        fun newInstance(bookId: Long) = BookPlayFragment().apply {
+            arguments = Bundle().apply {
+                putLong(NI_BOOK_ID, bookId)
+            }
         }
     }
 }