+++ audiobook/src/main/kotlin/de/ph1b/audiobook/uitools/ThemeUtil.kt
@@ -7,9 +7,58 @@ import android.os.Build
 import android.support.annotation.*
 import android.support.v4.content.ContextCompat
 import android.support.v4.graphics.drawable.DrawableCompat
+import android.view.View
 import android.widget.EditText
 import android.widget.NumberPicker
 import de.ph1b.audiobook.R
+
+fun NumberPicker.theme() {
+    if (Build.VERSION.SDK_INT < Build.VERSION_CODES.LOLLIPOP) {
+        val colorAccent = ContextCompat.getColor(context, R.color.accent)
+        var i = 0
+        val count = childCount
+        while (i < count) {
+            val child = getChildAt(i)
+            try {
+                val selectorWheelPaintField = javaClass.getDeclaredField("mSelectorWheelPaint")
+                selectorWheelPaintField.isAccessible = true
+                (selectorWheelPaintField.get(this) as Paint).color = colorAccent
+                (child as EditText?)?.setTextColor(colorAccent)
+                invalidate()
+            } catch (e: NoSuchFieldException) {
+                e.printStackTrace()
+            } catch (e: IllegalAccessException) {
+                e.printStackTrace()
+            } catch (e: IllegalArgumentException) {
+                e.printStackTrace()
+            }
+
+            try {
+                val f1 = Class.forName("android.widget.NumberPicker").getDeclaredField("mSelectionDivider")
+                f1.isAccessible = true
+                val dividerDrawable = DrawableCompat.wrap(f1.get(this) as Drawable)
+                DrawableCompat.setTint(dividerDrawable, colorAccent)
+            } catch (e: Exception) {
+                e.printStackTrace()
+            }
+
+            i++
+        }
+        invalidate()
+    }
+}
+
+fun View.setVisible() {
+    visibility = View.VISIBLE
+}
+
+fun View.setInvisible() {
+    visibility = View.INVISIBLE
+}
+
+fun View.setGone() {
+    visibility = View.GONE
+}
 
 object ThemeUtil {
 
@@ -22,42 +71,6 @@ object ThemeUtil {
             throw IllegalArgumentException("Resource with attr=$attr not found")
         }
         return resId
-    }
-
-    fun theme(numberPicker: NumberPicker) {
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.LOLLIPOP) {
-            val colorAccent = ContextCompat.getColor(numberPicker.context, R.color.accent)
-            var i = 0
-            val count = numberPicker.childCount
-            while (i < count) {
-                val child = numberPicker.getChildAt(i)
-                try {
-                    val selectorWheelPaintField = numberPicker.javaClass.getDeclaredField("mSelectorWheelPaint")
-                    selectorWheelPaintField.isAccessible = true
-                    (selectorWheelPaintField.get(numberPicker) as Paint).color = colorAccent
-                    (child as EditText).setTextColor(colorAccent)
-                    numberPicker.invalidate()
-                } catch (e: NoSuchFieldException) {
-                    e.printStackTrace()
-                } catch (e: IllegalAccessException) {
-                    e.printStackTrace()
-                } catch (e: IllegalArgumentException) {
-                    e.printStackTrace()
-                }
-
-                try {
-                    val f1 = Class.forName("android.widget.NumberPicker").getDeclaredField("mSelectionDivider")
-                    f1.isAccessible = true
-                    val dividerDrawable = DrawableCompat.wrap(f1.get(numberPicker) as Drawable)
-                    DrawableCompat.setTint(dividerDrawable, colorAccent)
-                } catch (e: Exception) {
-                    e.printStackTrace()
-                }
-
-                i++
-            }
-            numberPicker.invalidate()
-        }
     }
 
     enum class Theme internal constructor(@StyleRes val themeId: Int, @StringRes val nameId: Int, @ColorRes val colorId: Int) {